+++ src/main/java/com/epam/reportportal/cucumber/AbstractReporter.java
@@ -28,6 +28,7 @@ import cucumber.api.*;
 import cucumber.api.event.*;
 import cucumber.api.formatter.Formatter;
 import io.reactivex.Maybe;
+import org.apache.commons.lang3.tuple.Pair;
 import org.apache.tika.mime.MimeTypeException;
 import org.apache.tika.mime.MimeTypes;
 import org.slf4j.Logger;
@@ -37,7 +38,11 @@ import rp.com.google.common.base.Suppliers;
 
 import java.util.Calendar;
 import java.util.Date;
+import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
 
+import static com.epam.reportportal.cucumber.Utils.getCodeRef;
+import static com.epam.reportportal.cucumber.Utils.getDescription;
 import static rp.com.google.common.base.Strings.isNullOrEmpty;
 
 /**
@@ -57,22 +62,19 @@ public abstract class AbstractReporter implements Formatter {
 	protected static final String COLON_INFIX = ": ";
 
 	protected static final String SKIPPED_ISSUE_KEY = "skippedIssue";
+	protected Supplier<Launch> launch;
 
-	/* feature context */
-	protected RunningContext.FeatureContext currentFeatureContext;
+	private final Map<String, RunningContext.FeatureContext> currentFeatureContextMap = new ConcurrentHashMap<>();
 
-	/* scenario context */
-	protected RunningContext.ScenarioContext currentScenarioContext;
+	private final Map<Pair<String, String>, RunningContext.ScenarioContext> currentScenarioContextMap = new ConcurrentHashMap<>();
 
-	protected Supplier<Launch> RP;
+	private final Map<Long, RunningContext.ScenarioContext> threadCurrentScenarioContextMap = new ConcurrentHashMap<>();
 
-	protected void setCurrentFeatureContext(RunningContext.FeatureContext currentFeatureContext) {
-		this.currentFeatureContext = currentFeatureContext;
-	}
+	// There is no event for recognizing end of feature in Cucumber.
+	// This map is used to record the last scenario time and its feature uri.
+	// End of feature occurs once launch is finished.
+	private final Map<String, Date> featureEndTime = new ConcurrentHashMap<>();
 
-	protected void setCurrentScenarioContext(RunningContext.ScenarioContext currentScenarioContext) {
-		this.currentScenarioContext = currentScenarioContext;
-	}
 
 	/**
 	 * Registers an event handler for a specific event.
@@ -108,7 +110,7 @@ public abstract class AbstractReporter implements Formatter {
 	 */
 	protected void beforeLaunch() {
 		startLaunch();
-		RP.get().start();
+		launch.get().start();
 	}
 
 	/**
@@ -126,38 +128,26 @@ public abstract class AbstractReporter implements Formatter {
 	protected void afterLaunch() {
 		FinishExecutionRQ finishLaunchRq = new FinishExecutionRQ();
 		finishLaunchRq.setEndTime(Calendar.getInstance().getTime());
-		RP.get().finish(finishLaunchRq);
-	}
-
-	/**
-	 * Manipulations before the feature starts
-	 */
-	protected void beforeFeature() {
-		System.out.println("beforeFeature()");
-		startFeature();
-	}
-
-	/**
-	 * Finish Cucumber feature
-	 */
-	protected void afterFeature() {
-		Utils.finishTestItem(RP.get(), currentFeatureContext.getFeatureId());
-		currentFeatureContext = null;
+		launch.get().finish(finishLaunchRq);
 	}
 
 	/**
 	 * Start Cucumber scenario
 	 */
-	protected void beforeScenario() {
+	protected void beforeScenario(RunningContext.FeatureContext currentFeatureContext,
+			RunningContext.ScenarioContext currentScenarioContext, String scenarioName) {
+		String description = getDescription(currentFeatureContext.getUri());
+		String codeRef = getCodeRef(currentFeatureContext.getUri(), currentScenarioContext.getLine());
 		Maybe<String> id = Utils.startNonLeafNode(
-				RP.get(),
+				launch.get(),
 				currentFeatureContext.getFeatureId(),
 				Utils.buildNodeName(currentScenarioContext.getKeyword(),
 						AbstractReporter.COLON_INFIX,
 						currentScenarioContext.getName(),
 						currentScenarioContext.getOutlineIteration()
 				),
-				currentFeatureContext.getUri() + ":" + currentScenarioContext.getLine(),
+				description,
+				codeRef,
 				currentScenarioContext.getAttributes(),
 				getScenarioTestItemType()
 		);
@@ -166,32 +156,18 @@ public abstract class AbstractReporter implements Formatter {
 
 	/**
 	 * Finish Cucumber scenario
+	 * Put scenario end time in a map to check last scenario end time per feature
 	 */
 	protected void afterScenario(TestCaseFinished event) {
-		Utils.finishTestItem(RP.get(), currentScenarioContext.getId(), event.result.getStatus());
-		currentScenarioContext = null;
-	}
-
-	/**
-	 * Start Cucumber feature
-	 */
-	protected void startFeature() {
-		StartTestItemRQ rq = new StartTestItemRQ();
-		Maybe<String> root = getRootItemId();
-		rq.setDescription(currentFeatureContext.getUri());
-		rq.setName(Utils.buildNodeName(
-				currentFeatureContext.getFeature().getKeyword(),
-				AbstractReporter.COLON_INFIX,
-				currentFeatureContext.getFeature().getName(),
-				null
-		));
-		rq.setAttributes(currentFeatureContext.getAttributes());
-		rq.setStartTime(Calendar.getInstance().getTime());
-		rq.setType(getFeatureTestItemType());
-		if (null == root) {
-			currentFeatureContext.setFeatureId(RP.get().startTestItem(rq));
-		} else {
-			currentFeatureContext.setFeatureId(RP.get().startTestItem(root, rq));
+		RunningContext.ScenarioContext currentScenarioContext = getCurrentScenarioContext();
+		for (Map.Entry<Pair<String, String>, RunningContext.ScenarioContext> scenarioContext : currentScenarioContextMap.entrySet()) {
+			if (scenarioContext.getValue().getLine() == currentScenarioContext.getLine()) {
+				currentScenarioContextMap.remove(scenarioContext.getKey());
+				Date endTime = Utils.finishTestItem(launch.get(), currentScenarioContext.getId(), event.result.getStatus());
+				String featureURI = scenarioContext.getKey().getValue();
+				featureEndTime.put(featureURI, endTime);
+				break;
+			}
 		}
 	}
 
@@ -199,7 +175,7 @@ public abstract class AbstractReporter implements Formatter {
 	 * Start RP launch
 	 */
 	protected void startLaunch() {
-		RP = Suppliers.memoize(new Supplier<Launch>() {
+		launch = Suppliers.memoize(new Supplier<Launch>() {
 
 			/* should no be lazy */
 			private final Date startTime = Calendar.getInstance().getTime();
@@ -326,6 +302,29 @@ public abstract class AbstractReporter implements Formatter {
 
 	protected abstract Maybe<String> getRootItemId();
 
+	protected RunningContext.ScenarioContext getCurrentScenarioContext() {
+		return threadCurrentScenarioContextMap.get(Thread.currentThread().getId());
+	}
+
+	private RunningContext.FeatureContext createFeatureContext(TestCase testCase, String featureURI) {
+		RunningContext.FeatureContext currentFeatureContext;
+		currentFeatureContext = new RunningContext.FeatureContext().processTestSourceReadEvent(testCase);
+		currentFeatureContextMap.put(featureURI, currentFeatureContext);
+		String featureKeyword = currentFeatureContext.getFeature().getKeyword();
+		String featureName = currentFeatureContext.getFeature().getName();
+
+		StartTestItemRQ rq = new StartTestItemRQ();
+		Maybe<String> root = getRootItemId();
+		rq.setDescription(getDescription(currentFeatureContext.getUri()));
+		rq.setCodeRef(getCodeRef(currentFeatureContext.getUri(), 0));
+		rq.setName(Utils.buildNodeName(featureKeyword, AbstractReporter.COLON_INFIX, featureName, null));
+		rq.setAttributes(currentFeatureContext.getAttributes());
+		rq.setStartTime(Calendar.getInstance().getTime());
+		rq.setType(getFeatureTestItemType());
+		currentFeatureContext.setFeatureId(root == null ? launch.get().startTestItem(rq) : launch.get().startTestItem(root, rq));
+		return currentFeatureContext;
+	}
+
 	/**
 	 * Private part that responsible for handling events
 	 */
@@ -388,9 +387,7 @@ public abstract class AbstractReporter implements Formatter {
 		return new EventHandler<TestRunFinished>() {
 			@Override
 			public void receive(TestRunFinished event) {
-				if (currentFeatureContext != null) {
-					handleEndOfFeature();
-				}
+				handleEndOfFeature();
 				afterLaunch();
 			}
 		};
@@ -414,39 +411,53 @@ public abstract class AbstractReporter implements Formatter {
 		};
 	}
 
-	private void handleStartOfFeature(TestCase testCase) {
-		currentFeatureContext = new RunningContext.FeatureContext().processTestSourceReadEvent(testCase);
-		beforeFeature();
-	}
-
 	private void handleEndOfFeature() {
-		afterFeature();
+		for (RunningContext.FeatureContext value : currentFeatureContextMap.values()) {
+			Date featureCompletionDateTime = featureEndTime.get(value.getUri());
+			Utils.finishFeature(launch.get(), value.getFeatureId(), featureCompletionDateTime);
+		}
+		currentFeatureContextMap.clear();
 	}
 
 	private void handleStartOfTestCase(TestCaseStarted event) {
 		TestCase testCase = event.testCase;
-		if (currentFeatureContext != null && !testCase.getUri().equals(currentFeatureContext.getUri())) {
-			handleEndOfFeature();
-		}
-		if (currentFeatureContext == null) {
-			handleStartOfFeature(testCase);
-		}
+		RunningContext.FeatureContext featureContext = new RunningContext.FeatureContext().processTestSourceReadEvent(testCase);
+		String featureUri = featureContext.getUri();
+		RunningContext.FeatureContext currentFeatureContext = currentFeatureContextMap.get(featureUri);
+
+		currentFeatureContext = currentFeatureContext == null ? createFeatureContext(testCase, featureUri) : currentFeatureContext;
+
 		if (!currentFeatureContext.getUri().equals(testCase.getUri())) {
 			throw new IllegalStateException("Scenario URI does not match Feature URI.");
 		}
+
+		RunningContext.ScenarioContext scenarioContext = currentFeatureContext.getScenarioContext(testCase);
+		String scenarioName = Utils.buildNodeName(
+				scenarioContext.getKeyword(),
+				AbstractReporter.COLON_INFIX,
+				scenarioContext.getName(),
+				scenarioContext.getOutlineIteration()
+		);
+
+		Pair<String, String> scenarioNameFeatureURI = Pair.of(testCase.getScenarioDesignation(), currentFeatureContext.getUri());
+		RunningContext.ScenarioContext currentScenarioContext = currentScenarioContextMap.get(scenarioNameFeatureURI);
+
 		if (currentScenarioContext == null) {
 			currentScenarioContext = currentFeatureContext.getScenarioContext(testCase);
+			currentScenarioContextMap.put(scenarioNameFeatureURI, currentScenarioContext);
+			threadCurrentScenarioContextMap.put(Thread.currentThread().getId(), currentScenarioContext);
 		}
-		beforeScenario();
+
+		beforeScenario(currentFeatureContext, currentScenarioContext, scenarioName);
 	}
 
 	private void handleTestStepStarted(TestStepStarted event) {
 		TestStep testStep = event.testStep;
 		if (testStep instanceof HookTestStep) {
-			beforeHooks(testStep.getHookType());
+			beforeHooks(((HookTestStep)testStep).getHookType());
 		} else {
-			if (currentScenarioContext.withBackground()) {
-				currentScenarioContext.nextBackgroundStep();
+			if (getCurrentScenarioContext().withBackground()) {
+				getCurrentScenarioContext().nextBackgroundStep();
 			}
 			beforeStep(testStep);
 		}