+++ src/main/java/com/epam/reportportal/cucumber/ScenarioReporter.java
@@ -26,6 +26,8 @@ import rp.com.google.common.base.Suppliers;
 
 import java.util.Calendar;
 
+import static cucumber.api.Result.Type.PASSED;
+
 /**
  * Cucumber reporter for ReportPortal that reports scenarios as test methods.
  * <p>
@@ -49,6 +51,9 @@ import java.util.Calendar;
  */
 public class ScenarioReporter extends AbstractReporter {
 	private static final String SEPARATOR = "-------------------------";
+	private static final String EMPTY_SUFFIX = "";
+	private static final String INFO = "INFO";
+	private static final String STEP_ = "STEP ";
 
 	protected Supplier<Maybe<String>> rootSuiteId;
 
@@ -60,20 +65,23 @@ public class ScenarioReporter extends AbstractReporter {
 
 	@Override
 	protected void beforeStep(TestStep testStep) {
+		RunningContext.ScenarioContext currentScenarioContext = getCurrentScenarioContext();
 		Step step = currentScenarioContext.getStep(testStep);
-		String decoratedStepName = decorateMessage(Utils.buildNodeName(
-				currentScenarioContext.getStepPrefix(),
+		int lineInFeaturefile = step.getLocation().getLine();
+		String decoratedStepName = lineInFeaturefile + decorateMessage(Utils.buildNodeName(currentScenarioContext.getStepPrefix(),
 				step.getKeyword(),
 				Utils.getStepName(testStep),
-				" "
+				EMPTY_SUFFIX
 		));
 		String multilineArg = Utils.buildMultilineArgument(testStep);
-		Utils.sendLog(decoratedStepName + multilineArg, "INFO", null);
+		Utils.sendLog(decoratedStepName + multilineArg, INFO, null);
 	}
 
 	@Override
 	protected void afterStep(Result result) {
-		reportResult(result, decorateMessage("STEP " + result.getStatus().toString().toUpperCase()));
+		if (!result.is(PASSED)) {
+			reportResult(result, decorateMessage(STEP_ + result.getStatus().toString().toUpperCase()));
+		}
 	}
 
 	@Override
@@ -116,7 +124,7 @@ public class ScenarioReporter extends AbstractReporter {
 	 * Start root suite
 	 */
 	protected void finishRootItem() {
-		Utils.finishTestItem(RP.get(), rootSuiteId.get());
+		Utils.finishTestItem(launch.get(), rootSuiteId.get());
 		rootSuiteId = null;
 	}
 
@@ -124,15 +132,12 @@ public class ScenarioReporter extends AbstractReporter {
 	 * Start root suite
 	 */
 	protected void startRootItem() {
-		rootSuiteId = Suppliers.memoize(new Supplier<Maybe<String>>() {
-			@Override
-			public Maybe<String> get() {
-				StartTestItemRQ rq = new StartTestItemRQ();
-				rq.setName("Root User Story");
-				rq.setStartTime(Calendar.getInstance().getTime());
-				rq.setType("STORY");
-				return RP.get().startTestItem(rq);
-			}
+		rootSuiteId = Suppliers.memoize(() -> {
+			StartTestItemRQ rq = new StartTestItemRQ();
+			rq.setName("Root User Story");
+			rq.setStartTime(Calendar.getInstance().getTime());
+			rq.setType("STORY");
+			return launch.get().startTestItem(rq);
 		});
 	}